[#64] move marker to place folder
