--- app/modules/google-sdk/shared/place.ts
@@ -1,3 +1,0 @@
-/**
- * Created by thomasmichelet on 15/09/2017.
- */
